python_operator/requirements.txt
certifi
charset-normalizer
idna
munch
pykube-ng
PyYAML
requests
urllib3
